style(Header HP): manage auth button and manage mobile style
